Conflicts:
	AppController.m
	BrowserView.h
	Database.m
	HelperFunctions.m
	Preferences.m

+++ ArticleController.m
@@ -371,7 +371,7 @@
 		Article * article = [filteredArray objectAtIndex:index];
 		if (([article folderId] == folderIdOfArticleToPreserve) && [[article guid] isEqualToString:guidOfArticleToPreserve])
 			guidOfArticleToPreserve = @"";
-		else if ((comparator != nil) && !((BOOL)(int)[ArticleFilter performSelector:comparator withObject:article]))
+		else if ((comparator != nil) && !((BOOL)(NSInteger)[ArticleFilter performSelector:comparator withObject:article]))
 			[filteredArray removeObjectAtIndex:index];
 	}
